+++ x/icamauth/README.md
@@ -1,0 +1,4 @@
+# ICAMauth module
+
+Interchain account message authentication module (ICAMauth module) is copied from the [interchain-accounts-demo](https://github.com/cosmos/interchain-accounts-demo) respository
+with minor modifications and refator(rename intertx to icamauth). All credits and big thanks go to the original authors.
