--- x/inter-tx/README.md
@@ -1,4 +1,0 @@
-# Mauth module
-
-Mauth module is copied from the [interchain-accounts-demo](https://github.com/cosmos/interchain-accounts-demo) respository
-with minor modifications. All credits and big thanks go to the original authors.
